refactor(contract): move the state-patching helper into tee::test_utils
The crate convention is a test_utils submodule per domain module, and the
helper edits the launcher allowlist, which lives under tee. Also drops the
round-trip assert: borsh encoding is canonical, so deserializing and
reserializing the same type is byte-identical by construction.
