+++ crates/contract/src/tee/test_utils.rs
@@ -3,11 +3,13 @@
 //! This module provides helper functions and types for testing TEE state,
 //! attestation behavior, and general contract state management.
 
+use crate::MpcContract;
 use crate::primitives::test_utils::{gen_account_id, gen_seed};
 use crate::tee::{measurements::ContractExpectedMeasurements, tee_state::TeeState};
 use mpc_attestation::attestation::default_measurements;
-use mpc_primitives::hash::{LauncherImageHash, NodeImageHash};
+use mpc_primitives::hash::{LauncherDockerComposeHash, LauncherImageHash, NodeImageHash};
 use near_account_id::AccountId;
+use near_sdk::borsh::{self, BorshDeserialize};
 use near_sdk::{BlockHeight, NearToken, PublicKey, test_utils::VMContextBuilder, testing_env};
 use rand::Rng;
 use std::time::Duration;
@@ -114,4 +116,20 @@ pub fn whitelist_dstack_measurements(
     for &measurements in default_measurements() {
         tee_state.add_measurement(ContractExpectedMeasurements::from(measurements));
     }
+}
+
+/// Adds `compose_hash` to `launcher_hash`'s allowlist entry in a raw `STATE` blob, for
+/// sandbox tests to patch back in. The attestation fixture's compose hash is not
+/// derivable from the compiled-in template, so no vote can allow it.
+pub fn allow_launcher_compose_hash_in_state(
+    state: &[u8],
+    launcher_hash: &LauncherImageHash,
+    compose_hash: LauncherDockerComposeHash,
+) -> Vec<u8> {
+    let mut contract = MpcContract::try_from_slice(state).expect("STATE deserializes");
+    contract
+        .tee_state
+        .allowed_launcher_images
+        .allow_compose_hash(launcher_hash, compose_hash);
+    borsh::to_vec(&contract).expect("STATE serializes")
 }
